+++ command/base/base.go
@@ -1,0 +1,7 @@
+package base
+
+// Command provides a base structure for all commands, containing the common
+// command line options like the repository URL, the tag to use, and all
+// the authentication-related options.
+type Command struct {
+}